Add files via upload
In this project, I worked on downloading a database and restoring it on the server. I had then query the database to get customer details like name, phone number, email ID, sales made in a particular month, increase in month-on-month sales and even the total sales made to a particular customer.The objective to is understand how to query large database.